[24916] File uploads through clipboard paste
This updates ng-file-upload and allows pasting images in the formattable
description field.

https://community.openproject.com/projects/openproject/work_packages/24916/activity
