+++ app/assets/stylesheets/content/work_packages/inplace_editing/_edit_fields.sass
@@ -164,3 +164,11 @@
 // Allow horizontal scrolling in descripton field
 .work-packages--details--description .wp-table--cell-span .inplace-edit--read-value--value-span
   overflow-x: auto
+
+// Style the attachment hint label of the description field
+.wp-edit-field-attachment-label
+  position: absolute
+  left: 1px
+  font-size: 0.8rem
+  padding-top: 5px
+  font-style: italic
