+++ config/locales/js-en.yml
@@ -208,6 +208,7 @@ en:
     label_drop_files: Drop files here
     label_drop_files_hint: or click to add files
     label_add_attachments: "Add attachments"
+    label_formattable_attachment_hint: "Attach and link files by dropping on this field, or pasting from the clipboard."
     label_remove_file: "Delete %{fileName}"
     label_remove_watcher: "Remove watcher %{name}"
     label_remove_all_files: Delete all files
