+++ frontend/app/components/work-packages/wp-attachments-formattable-field/wp-attachments-formattable.enums.ts
@@ -1,14 +1,11 @@
 export enum InsertMode {
-    ATTACHMENT,
-    DELAYED_ATTACHMENT,
-    INLINE,
-    LINK
+  ATTACHMENT,
+  DELAYED_ATTACHMENT,
+  INLINE,
+  LINK
 }
 
 export enum ViewMode {
-    EDIT,
-    SHOW,
-    CREATE
+  EDIT,
+  SHOW
 }
-
-
