+++ frontend/app/components/work-packages/wp-attachments-formattable-field/wp-attachments-formattable.interfaces.ts
@@ -1,9 +1,9 @@
 import {InsertMode} from './wp-attachments-formattable.enums';
 
 export interface IApplyAttachmentMarkup {
-    contentToInsert: string;
+  contentToInsert:string;
 
-    insertAttachmentLink: (url: string, insertMode: InsertMode, addLineBreak?:boolean) => void;
-    insertWebLink: (url: string, insertMode: InsertMode) => void;
-    save: () => void;
+  insertAttachmentLink:(url:string, insertMode:InsertMode, addLineBreak?:boolean) => void;
+  insertWebLink:(url:string, insertMode:InsertMode) => void;
+  save:() => void;
 }
