+++ frontend/app/components/wp-attachments/wp-attachments-upload/wp-attachments-upload.directive.test.ts
@@ -113,10 +113,6 @@ describe('wpAttachmentsUpload directive', () => {
       expect(rootElement).to.have.length(1);
     });
 
-    it('should set the max size property of the element to the configured value', () => {
-      expect(rootElement.attr('ngf-max-size')).to.equal(mockMaxSize.toString());
-    });
-
     describe('when clicking the parent element', () => {
       var clicked:any;
