+++ frontend/package.json
@@ -52,13 +52,13 @@
     "angular-animate": "~1.5.9",
     "angular-aria": "~1.5.9",
     "angular-cache": "~4.6.0",
-    "angular-context-menu": "opf/angular-context-menu#a908eccaec323cd66973d58af4965694bdff16a1",
+    "angular-context-menu": "git://github.com/opf/angular-context-menu#a908eccaec323cd66973d58af4965694bdff16a1",
     "angular-dragula": "~1.2.8",
     "angular-elastic": "2.5.0",
     "angular-i18n": "~1.3.0",
-    "angular-modal": "finnlabs/angular-modal#d45eb9ceb720b8785613ba89ba0f14f8ab197569",
+    "angular-modal": "git://github.com/finnlabs/angular-modal#d45eb9ceb720b8785613ba89ba0f14f8ab197569",
     "angular-sanitize": "~1.3.14",
-    "angular-truncate": "sparkalow/angular-truncate#fdf60fda265042d12e9414b5354b2cc52f1419de",
+    "angular-truncate": "git://github.com/sparkalow/angular-truncate#fdf60fda265042d12e9414b5354b2cc52f1419de",
     "angular-ui-bootstrap": "^2.2.0",
     "angular-ui-router": "^1.0.0-rc.1",
     "atoa": "^1.0.0",
@@ -94,7 +94,7 @@
     "mousetrap": "~1.6.0",
     "ng-annotate-loader": "^0.2.0",
     "ng-dialog": "^0.6.4",
-    "ng-file-upload": "~5.0.9",
+    "ng-file-upload": "^12.2.13",
     "ngtemplate-loader": "^0.1.2",
     "observable-array": "0.0.4",
     "phantomjs-polyfill": "0.0.2",
